+++ src/lib/copy-text.ts
@@ -1,3 +1,5 @@
+import { updateTooltipText } from '@lib/tooltip'
+
 export const copyTextClasses = 'copyclip break-line use-tooltip'
 export const copyTextMsg = "cliquer pour copier l'élement"
 const copyTextMsgCopied = 'copié dans le presse-papier !'
@@ -11,9 +13,7 @@ export function copyTextListenClick() {
       return
     const text = target.textContent.trim()
     await navigator.clipboard.writeText(text)
-    const tooltip = document.getElementById('powerTip')
-    if (!tooltip) return
-    tooltip.textContent = copyTextMsgCopied
-    setTimeout(() => (tooltip.textContent = copyTextMsg), 1000)
+    updateTooltipText(copyTextMsgCopied)
+    setTimeout(() => updateTooltipText(copyTextMsg), 1000)
   })
 }